HawkerSG/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
email-validator
passlib
argon2-cffi
python-dotenv
sendgrid
python-jose[cryptography]
python-multipart
pandas
openpyxl
openai
requests
pyinstrument
pytest
pytest-asyncio
httpx